requirements-light.txt
Flask==3.1.1
Werkzeug==3.1.3
pytest==8.4.1
pytest-xdist==3.8.0
gunicorn==23.0.0
# Basic file processing dependencies
requests==2.32.4
pandas==2.2.3
# Vector DB - lighter alternative
faiss-cpu==1.9.0
numpy==2.3.1
